14030718- Refactor: Remove gRPC and ProtoBuf, add AutoMapper #26
