+++ Avhrm.Core/ApplicationServices.cs
@@ -1,0 +1,13 @@
+﻿using Microsoft.Extensions.DependencyInjection;
+
+namespace Avhrm.Application;
+public static class CoreServices
+{
+    public static void AddApplicationServices(this IServiceCollection services)
+    {
+        services.AddAutoMapper(options =>
+        {
+            options.AddProfile<ApplicationProfile>();
+        });
+    }
+}
