+++ Avhrm.Core/Avhrm.Application.csproj
@@ -7,20 +7,19 @@
 	</PropertyGroup>
 
 	<ItemGroup>
-		<PackageReference Include="Google.Protobuf" Version="3.25.0" />
-		<PackageReference Include="Grpc.Core" Version="2.46.6" />
-		<PackageReference Include="Grpc.Tools" Version="2.5.0">
-			<PrivateAssets>all</PrivateAssets>
-			<IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
-		</PackageReference>
 		<PackageReference Include="Microsoft.EntityFrameworkCore" Version="8.0.0" />
 		<PackageReference Include="Microsoft.Extensions.DependencyInjection.Abstractions" Version="8.0.0" />
-		<PackageReference Include="protobuf-net.Grpc" Version="1.1.1" />
-		<PackageReference Include="protobuf-net.BuildTools" Version="3.2.33">
-			<PrivateAssets>all</PrivateAssets>
-			<IncludeAssets>runtime; build; native; contentfiles; analyzers</IncludeAssets>
-		</PackageReference>
-		<PackageReference Include="Grpc.Net.Client" Version="2.59.0" />
 	</ItemGroup>
 
+	<ItemGroup>
+		<PackageReference Include="AutoMapper" Version="13.0.1" />
+	</ItemGroup>
+
+	<ItemGroup>
+	  <ProjectReference Include="..\Avhrm.Domains\Avhrm.Domains.csproj" />
+	</ItemGroup>
+
+	<ItemGroup>
+		<Using Include="System.ComponentModel.DataAnnotations"/>
+	</ItemGroup>
 </Project>
