+++ Avhrm.Core/Common/BaseDto.cs
@@ -1,10 +1,6 @@
-﻿using ProtoBuf;
+﻿namespace Avhrm.Application;
 
-namespace Avhrm.Application.Common;
-
-[ProtoContract]
 public class BaseDto<T>
 {
-    [ProtoMember(1)]
     public T Value { get; set; }
 }
