+++ Avhrm.Core/Contracts/IAuthenticationService.cs
@@ -1,11 +1,6 @@
-﻿using Avhrm.Application.Common;
-using Avhrm.Application.Features.Account.Query.GerUserLogin;
-using ProtoBuf.Grpc;
-using ProtoBuf.Grpc.Configuration;
+﻿using ProtoBuf.Grpc;
 
 namespace Avhrm.Application;
-
-[Service]
 public interface IAuthenticationService
 {
     Task<BaseDto<string>> Authenticate(GetUserLoginQuery request, CallContext callContext = default);
