--- Avhrm.Core/Contracts/IVacationRequest.cs
@@ -1,19 +1,0 @@
-﻿using Avhrm.Application.Common;
-using ProtoBuf.Grpc;
-using ProtoBuf.Grpc.Configuration;
-
-namespace Avhrm.Application;
-
-[Service]
-public interface IVacationRequest
-{
-    //Task<BaseDto<bool>> InsertVacationRequest(VacationRequest request, CallContext context = default);
-
-    //Task<List<VacationRequest>> GetVacationRequests(CallContext context = default);
-
-    //Task<VacationRequest> GetVacationRequestById(BaseDto<int>id, CallContext context = default);
-
-    //Task<BaseDto<bool>> UpdateVacationRequest(VacationRequest request, CallContext context = default);
-    
-    //Task<BaseDto<bool>> DeleteVacationRequest(VacationRequest request, CallContext context = default);
-}
