--- Avhrm.Core/Contracts/IWorkChallengeService.cs
@@ -1,11 +1,0 @@
-﻿using Avhrm.Application.Features.WorkChallenge.Query.GetAllWorkChallenge;
-using ProtoBuf.Grpc;
-using ProtoBuf.Grpc.Configuration;
-
-namespace Avhrm.Application;
-
-[Service]
-public interface IWorkChallengeService
-{
-    Task<List<GetAllWorkChallengeVm>> GetWorkChallengesByDepartmentId(CallContext context = default);
-}
