--- Avhrm.Core/CoreServices.cs
@@ -1,10 +1,0 @@
-﻿using Microsoft.Extensions.DependencyInjection;
-
-namespace Avhrm.Application;
-public static class CoreServices
-{
-    public static void AddCoreServices(this IServiceCollection services)
-    {
-       
-    }
-}
