+++ Avhrm.Core/Features/Customer/Queries/GetAllCustomers/GetAllCustomersVm.cs
@@ -1,0 +1,7 @@
+﻿namespace Avhrm.Application.Features;
+
+public class GetAllCustomersVm
+{
+    public int? Id { get; set; }
+    public string Name { get; set; }
+}
