+++ Avhrm.Core/Features/Project/Queries/GetAllProjects/GetAllProjectsVm.cs
@@ -1,0 +1,6 @@
+﻿namespace Avhrm.Application.Features;
+public class GetAllProjectsVm
+{
+    public int? Id { get; set; }
+    public string Name { get; set; }
+}
