+++ Avhrm.Core/Features/WorkChallenge/Queries/GetAllWorkChallenge/GetAllWorkChallengeVm.cs
@@ -1,0 +1,8 @@
+﻿namespace Avhrm.Application.Features;
+public class GetAllWorkChallengeVm
+{
+    public int Id { get; set; }
+    public string Description { get; set; }
+    public int DepartmentId { get; set; }
+    public string DepartmentName { get; set; }
+}
