+++ Avhrm.Core/Features/WorkType/Queries/GetAllWorkTypes/GetAllWorkTypesVm.cs
@@ -1,0 +1,7 @@
+﻿namespace Avhrm.Application.Features;
+public class GetAllWorkTypesVm
+{
+    public int Id { get; set; }
+    public string Description { get; set; }
+    public string DepartmentName { get; set; }
+}
