+++ Avhrm.Core/Features/WorkingReport/Commands/DeleteWorkReport/DeleteWorkReportCommand.cs
@@ -1,0 +1,5 @@
+﻿namespace Avhrm.Application.Features;
+public class DeleteWorkReportCommand
+{
+    public int? Id { get; set; }
+}
