--- Avhrm.Core/Features/WorkingReport/Enum/WorkDayType.cs
@@ -1,6 +1,0 @@
-﻿namespace Avhrm.Application.Features.WorkingReport.Enum;
-public enum WorkDayType
-{
-    WorkDay,
-    Holiday
-}
