--- Avhrm.Core/Features/WorkingReport/Enum/WorkReportTimeOfDay.cs
@@ -1,7 +1,0 @@
-﻿namespace Avhrm.Application.Features.WorkingReport.Enum;
-public enum WorkReportTimeOfDay
-{
-    Morning,
-    Noon,
-    Afternoon
-}
