+++ Avhrm.Core/Features/WorkingReport/Enums/WorkDayType.cs
@@ -1,0 +1,6 @@
+﻿namespace Avhrm.Application.Features;
+public enum WorkDayType
+{
+    WorkDay,
+    Holiday
+}
