+++ Avhrm.Core/Features/WorkingReport/Enums/WorkReportTimeOfDay.cs
@@ -1,0 +1,7 @@
+﻿namespace Avhrm.Application.Features;
+public enum WorkReportTimeOfDay
+{
+    Morning,
+    Noon,
+    Afternoon
+}
