+++ Avhrm.Core/Features/WorkingReport/Queries/GetUserWorkingReportByDate/GetUserWorkingReportByDateVm.cs
@@ -1,0 +1,9 @@
+﻿namespace Avhrm.Application.Features;
+public class GetUserWorkingReportByDateVm
+{
+    public int Id { get; set; }
+    public string PersianDate { get; set; }
+    public decimal SpentHours { get; set; }
+    public string WorkTypeDescription { get; set; }
+    public string Desc { get; set; }
+}
