+++ Avhrm.Core/Features/WorkingReport/Queries/GetWorkReportById/GetWorkReportByIdQuery.cs
@@ -1,0 +1,5 @@
+﻿namespace Avhrm.Application.Features;
+public class GetWorkReportByIdQuery
+{
+    public int Id { get; set; }
+}
