--- Avhrm.Core/Features/WorkingReport/Query/GetWorkReportById/GetWorkReportByIdQuery.cs
@@ -1,10 +1,0 @@
-﻿using ProtoBuf;
-
-namespace Avhrm.Application.Features.WorkingReport.Query.GetWorkReportById;
-
-[ProtoContract]
-public class GetWorkReportByIdQuery
-{
-    [ProtoMember(1)]
-    public int Id { get; set; }
-}
