+++ Avhrm.Domains/Avhrm.Domains.csproj
@@ -12,7 +12,7 @@
 
 	<ItemGroup>
 		<PackageReference Include="Microsoft.AspNetCore.Identity.EntityFrameworkCore" Version="8.0.0" />
-		<PackageReference Include="AutoMapper" Version="13.0.1" />
+		
 	</ItemGroup>
 
 	<ItemGroup>
